Change (#123): move functions to add task
The functions to add tasks to clusters was moved
out of the cluster-class in order to reduce the
size of the class.
