+++ src/core/Hanami/src/core/cluster/cluster.h
@@ -106,42 +106,6 @@ public:
 
     // task-handling
     void updateClusterState();
-    const std::string addImageLearnTask(const std::string &name,
-                                        const std::string &userId,
-                                        const std::string &projectId,
-                                        float* inputData,
-                                        const uint64_t numberOfInputsPerCycle,
-                                        const uint64_t numberOfOuputsPerCycle,
-                                        const uint64_t numberOfCycle);
-    const std::string addImageRequestTask(const std::string &name,
-                                          const std::string &userId,
-                                          const std::string &projectId,
-                                          float* inputData,
-                                          const uint64_t numberOfInputsPerCycle,
-                                          const uint64_t numberOfOuputsPerCycle,
-                                          const uint64_t numberOfCycle);
-    const std::string addTableLearnTask(const std::string &name,
-                                        const std::string &userId,
-                                        const std::string &projectId,
-                                        float* inputData,
-                                        float* outputData,
-                                        const uint64_t numberOfInputs,
-                                        const uint64_t numberOfOutputs,
-                                        const uint64_t numberOfCycle);
-    const std::string addTableRequestTask(const std::string &name,
-                                          const std::string &userId,
-                                          const std::string &projectId,
-                                          float* inputData,
-                                          const uint64_t numberOfInputs,
-                                          const uint64_t numberOfOutputs,
-                                          const uint64_t numberOfCycle);
-    const std::string addClusterSnapshotSaveTask(const std::string &snapshotName,
-                                                 const std::string &userId,
-                                                 const std::string &projectId);
-    const std::string addClusterSnapshotRestoreTask(const std::string &name,
-                                                    const std::string &snapshotInfo,
-                                                    const std::string &userId,
-                                                    const std::string &projectId);
 
     // tasks
     Task* getActualTask() const;
@@ -160,9 +124,10 @@ public:
     ClusterProcessingMode mode = NORMAL_MODE;
     HttpWebsocketThread* msgClient = nullptr;
 
+    Kitsunemimi::Statemachine* stateMachine = nullptr;
+    TaskHandle_State* taskHandleState = nullptr;
+
 private:
-    Kitsunemimi::Statemachine* m_stateMachine = nullptr;
-    TaskHandle_State* m_taskHandleState = nullptr;
     std::mutex m_segmentCounterLock;
 };
